Refactor: extract MainWindow into reusable UserControls
MainWindow.axaml drops from 536 to ~75 lines by extracting six reusable
components so the visual editor can open each file:

- Sidebar: nav + storage card
- LibraryPage: header, DataGrid/grid views, inspector, empty state
- LogoPicker: logo grid overlay (Reset/Custom/Close)
- TorrentPage: torrent pre-fill wizard
- SettingsPage: settings cards
- ActivityBar: status/progress/cancel/log toggle

Code-behind pointer handlers moved to their owning controls
(LibraryPage.OnGridItemClicked, LogoPicker.OnLogoClicked).
All 274 tests pass.
