+++ src/LincleLINK.App/Views/ActivityBar.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace LincleLINK.App.Views;
+
+public partial class ActivityBar : UserControl
+{
+    public ActivityBar()
+    {
+        InitializeComponent();
+    }
+}
