+++ src/LincleLINK.App/Views/MainWindow.axaml.cs
@@ -1,9 +1,6 @@
 using Avalonia.Controls;
-using Avalonia.Input;
-using LincleLINK.App.Logos;
 using LincleLINK.App.Services;
 using LincleLINK.App.ViewModels;
-using LincleLINK.Core.Domain;
 
 namespace LincleLINK.App.Views;
 
@@ -19,39 +16,11 @@ public partial class MainWindow : Window
         base.OnOpened(e);
         ThemeManager.ApplyTitleBar(this);
 
+        // Refresh after the window is shown and the dispatcher is pumping, so the
+        // initial instance list and status reach the (lazily realized) tab content.
         if (DataContext is MainViewModel viewModel)
         {
             _ = viewModel.InitializeAsync();
-        }
-    }
-
-    private void OnGridItemClicked(object? sender, PointerPressedEventArgs e)
-    {
-        if (sender is Control control && control.DataContext is InstanceListEntry entry)
-        {
-            if (DataContext is MainViewModel vm)
-            {
-                vm.SelectedInstance = entry;
-            }
-        }
-    }
-
-    private void OnLogoClicked(object? sender, PointerPressedEventArgs e)
-    {
-        if (sender is Control control && control.DataContext is LogoEntry logo)
-        {
-            if (DataContext is MainViewModel vm)
-            {
-                vm.SetCustomLogoCommand.Execute(logo);
-            }
-        }
-    }
-
-    private void OnCloseLogoPicker(object? sender, EventArgs e)
-    {
-        if (DataContext is MainViewModel vm)
-        {
-            vm.SetCustomLogoCommand.Execute(null);
         }
     }
 }
