+++ src/LincleLINK.App/Views/SettingsPage.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace LincleLINK.App.Views;
+
+public partial class SettingsPage : UserControl
+{
+    public SettingsPage()
+    {
+        InitializeComponent();
+    }
+}
