+++ src/LincleLINK.App/Views/Sidebar.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace LincleLINK.App.Views;
+
+public partial class Sidebar : UserControl
+{
+    public Sidebar()
+    {
+        InitializeComponent();
+    }
+}
