+++ src/LincleLINK.App/Views/TorrentPage.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace LincleLINK.App.Views;
+
+public partial class TorrentPage : UserControl
+{
+    public TorrentPage()
+    {
+        InitializeComponent();
+    }
+}
